+++ localtileserver/client.py
@@ -14,9 +14,10 @@ except ImportError:
     DatasetReaderBase = None
 
 from localtileserver.configure import get_default_client_params
-from localtileserver.server import ServerManager, launch_server
+from localtileserver.manager import AppManager
 from localtileserver.tileserver import get_building_docs, get_clean_filename, palette_valid_or_raise
 from localtileserver.utilities import add_query_parameters, save_file_from_request
+from server_thread import ServerManager, launch_server
 
 BUILDING_DOCS = get_building_docs()
 DEMO_REMOTE_TILE_SERVER = "https://tileserver.banesullivan.com/"
@@ -365,7 +366,8 @@ class TileClient(BaseTileClient):
         if DatasetReaderBase and isinstance(filename, DatasetReaderBase) and hasattr(filename, "name"):
             filename = filename.name
         super().__init__(filename)
-        self._key = launch_server(port, debug, host=host)
+        app = AppManager.get_or_create_app()
+        self._key = launch_server(app, port=port, debug=debug, host=host)
         # Store actual port just in case
         self._port = ServerManager.get_server(self._key).srv.port
         client_host, client_port, client_prefix = get_default_client_params(
@@ -376,6 +378,14 @@ class TileClient(BaseTileClient):
         self._client_prefix = client_prefix
         if BUILDING_DOCS and not client_host:
             self._client_host = DEMO_REMOTE_TILE_SERVER
+
+        if not debug:
+            logging.getLogger("gdal").setLevel(logging.ERROR)
+            logging.getLogger("large_image").setLevel(logging.ERROR)
+        else:
+            logging.getLogger("gdal").setLevel(logging.DEBUG)
+            logging.getLogger("large_image").setLevel(logging.DEBUG)
+            logging.getLogger("large_image_source_gdal").setLevel(logging.DEBUG)
 
     def shutdown(self, force: bool = False):
         if hasattr(self, "_key"):
